requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.9.7
djangorestframework==3.3.3

# Configuration
django-environ==0.4.0

# Forms
django-braces==1.9.0

# Models
django-model-utils==2.5

# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Time zones support
pytz==2016.4

# Redis support
django-redis==4.4.3
redis>=2.10.0

-e git+https://github.com/anthony-tresontani/django-adaptors.git@master#egg=adaptors
lxml==3.3.5
